requirements.txt
python-telegram-bot==20.0.0
selenium==4.22.2
beautifulsoup4==4.12.2
requests==2.31.0